+++ Simplest-YOLOv3-Pytorch/ImagesAndLabelsSet.py
@@ -19,6 +19,7 @@ import argparse
 import matplotlib.pyplot as plt
 from draw_box_utils import draw_box
 import json
+import sys
 
 
 img_formats = ['.bmp', '.jpg', '.jpeg', '.png', '.tif', '.dng']
@@ -365,7 +366,7 @@ def load_mosaic(self, index):
     # 随机旋转，缩放，平移以及错切
     img4, labels4 = random_affine(img4, labels4,
                                   degrees=0.,
-                                  translate=0.,
+                                  translate=0.1,
                                   scale=0.,
                                   shear=0.,
                                   border=-s // 2)  # border to remove
@@ -386,6 +387,7 @@ def load_mosaic2(self, index):
     :return:
     """
     # loads images in a mosaic
+    print("func: ",sys._getframe().f_code.co_name)
 
     labels4, segments4 = [], []
     s = self.img_size
@@ -429,136 +431,24 @@ def load_mosaic2(self, index):
 
     # Augment
     # 随机旋转，缩放，平移以及错切
-    img4, labels4 = random_affine(img4, labels4,
-                                degrees=0.,
-                                translate=0.25,
-                                scale=0.,
-                                shear=0.,
-                                border=-s // 2)  # border to remove
+    # img4, labels4 = random_affine(img4, labels4,
+    #                             degrees=0.,
+    #                             translate=0.,
+    #                             scale=0.,
+    #                             shear=0.,
+    #                             border=-s // 2)  # border to remove
+
+    img4, labels4 = random_perspective(img4, labels4, segments4,
+                                    degrees=0.,
+                                    translate=0.5+0.1,
+                                    scale=0.,
+                                    shear=0.,
+                                    perspective=0.,
+                                    border=(-256,-256))  # border to remove
 
 
     return img4, labels4
 
-def random_affine(img, targets=(), degrees=10, translate=.1, scale=.1, shear=10, border=0):
-    """随机旋转，缩放，平移以及错切"""
-    # torchvision.transforms.RandomAffine(degrees=(-10, 10), translate=(.1, .1), scale=(.9, 1.1), shear=(-10, 10))
-    # https://medium.com/uruvideo/dataset-augmentation-with-random-homographies-a8f4b44830d4
-    # targets = [cls, xyxy]
-
-    # 给定的输入图像的尺寸(416/512/640)，等于img4.shape / 2
-    height = img.shape[0] + border * 2
-    width = img.shape[1] + border * 2
-
-    # Rotation and Scale
-    # 生成旋转以及缩放矩阵
-    R = np.eye(3)  # 生成对角阵
-    a = random.uniform(-degrees, degrees)  # 随机旋转角度
-    s = random.uniform(1 - scale, 1 + scale)  # 随机缩放因子
-    R[:2] = cv2.getRotationMatrix2D(angle=a, center=(img.shape[1] / 2, img.shape[0] / 2), scale=s)
-
-    # Translation
-    # 生成平移矩阵
-    T = np.eye(3)
-    T[0, 2] = random.uniform(-translate, translate) * img.shape[0] + border  # x translation (pixels)
-    T[1, 2] = random.uniform(-translate, translate) * img.shape[1] + border  # y translation (pixels)
-
-    # Shear
-    # 生成错切矩阵
-    S = np.eye(3)
-    S[0, 1] = math.tan(random.uniform(-shear, shear) * math.pi / 180)  # x shear (deg)
-    S[1, 0] = math.tan(random.uniform(-shear, shear) * math.pi / 180)  # y shear (deg)
-
-    # Combined rotation matrix
-    M = S @ T @ R  # ORDER IS IMPORTANT HERE!!
-    if (border != 0) or (M != np.eye(3)).any():  # image changed
-        #依靠T进行平移，之后通过dsize只取了大图的左上角区域，实现了裁剪图片中间的目的
-        img = cv2.warpAffine(img, M[:2], dsize=(width, height), flags=cv2.INTER_LINEAR, borderValue=(114, 114, 114))
-
-    # Transform label coordinates
-    n = len(targets)
-    if n:
-        # warp points
-        xy = np.ones((n * 4, 3))
-        xy[:, :2] = targets[:, [1, 2, 3, 4, 1, 4, 3, 2]].reshape(n * 4, 2)  # x1y1, x2y2, x1y2, x2y1
-        # [4*n, 3] -> [n, 8]
-        xy = (xy @ M.T)[:, :2].reshape(n, 8)
-
-        # create new boxes
-        # 对transform后的bbox进行修正(假设变换后的bbox变成了菱形，此时要修正成矩形)
-        x = xy[:, [0, 2, 4, 6]]  # [n, 4]
-        y = xy[:, [1, 3, 5, 7]]  # [n, 4]
-        xy = np.concatenate((x.min(1), y.min(1), x.max(1), y.max(1))).reshape(4, n).T  # [n, 4]
-
-        # reject warped points outside of image
-        # 对坐标进行裁剪，防止越界
-        xy[:, [0, 2]] = xy[:, [0, 2]].clip(0, width)
-        xy[:, [1, 3]] = xy[:, [1, 3]].clip(0, height)
-        w = xy[:, 2] - xy[:, 0]
-        h = xy[:, 3] - xy[:, 1]
-
-        # 计算调整后的每个box的面积
-        area = w * h
-        # 计算调整前的每个box的面积
-        area0 = (targets[:, 3] - targets[:, 1]) * (targets[:, 4] - targets[:, 2])
-        # 计算每个box的比例
-        ar = np.maximum(w / (h + 1e-16), h / (w + 1e-16))  # aspect ratio
-        # 选取长宽大于4个像素，且调整前后面积比例大于0.2，且比例小于10的box
-        i = (w > 4) & (h > 4) & (area / (area0 * s + 1e-16) > 0.2) & (ar < 10)
-
-        targets = targets[i]
-        targets[:, 1:5] = xy[i]
-
-    return img, targets
-
-def letterbox(img: np.ndarray,
-              new_shape=(416, 416),
-              color=(114, 114, 114),
-              auto=True,
-              scale_fill=False,
-              scale_up=True):
-    """
-    将图片缩放调整到指定大小
-    :param img:
-    :param new_shape:
-    :param color:
-    :param auto:
-    :param scale_fill:
-    :param scale_up:
-    :return:
-    """
-
-    shape = img.shape[:2]  # [h, w]
-    if isinstance(new_shape, int):
-        new_shape = (new_shape, new_shape)
-
-    # scale ratio (new / old)
-    r = min(new_shape[0] / shape[0], new_shape[1] / shape[1])
-    if not scale_up:  # only scale down, do not scale up (for better test mAP) 对于大于指定输入大小的图片进行缩放,小于的不变
-        r = min(r, 1.0)
-
-    # compute padding
-    ratio = r, r  # width, height ratios
-    new_unpad = int(round(shape[1] * r)), int(round(shape[0] * r))
-    dw, dh = new_shape[1] - new_unpad[0], new_shape[0] - new_unpad[1]  # wh padding
-    if auto:  # minimun rectangle 保证原图比例不变，将图像最大边缩放到指定大小
-        # 这里的取余操作可以保证padding后的图片是32的整数倍
-        dw, dh = np.mod(dw, 32), np.mod(dh, 32)  # wh padding
-    elif scale_fill:  # stretch 简单粗暴的将图片缩放到指定尺寸
-        dw, dh = 0, 0
-        new_unpad = new_shape
-        ratio = new_shape[0] / shape[1], new_shape[1] / shape[0]  # wh ratios
-
-    dw /= 2  # divide padding into 2 sides 将padding分到上下，左右两侧
-    dh /= 2
-
-    # shape:[h, w]  new_unpad:[w, h]
-    if shape[::-1] != new_unpad:
-        img = cv2.resize(img, new_unpad, interpolation=cv2.INTER_LINEAR)
-    top, bottom = int(round(dh - 0.1)), int(round(dh + 0.1))  # 计算上下两侧的padding
-    left, right = int(round(dw - 0.1)), int(round(dw + 0.1))  # 计算左右两侧的padding
-
-    img = cv2.copyMakeBorder(img, top, bottom, left, right, cv2.BORDER_CONSTANT, value=color)  # add border
-    return img, ratio, (dw, dh)
 
 def imshow(img):
     #img = img / 2 + 0.5     # unnormalize
